feat(api): added logic to handle COUNTER 5.1 reports
also refactored COUNTER 5
